fix #201146 allow BUILD_LAME=FALSE
